ingest_mysql_data/requirements.txt
google-cloud-storage
pandas
PyMySQL
python-dotenv
cloud-sql-python-connector[pymysql]
